update tests style
