+++ byte_infer_perf/llm_perf/backends/ROCM/rocm_kernels/op_tests/test_common.py
@@ -6,13 +6,14 @@
 # @Email: lingpeng.jin@amd.com
 # @Create At: 2024-11-03 15:53:32
 # @Last Modified By: valarLip
-# @Last Modified At: 2024-11-03 16:20:14
+# @Last Modified At: 2024-11-08 23:44:57
 # @Description: This is description.
 
 import torch
 import numpy as np
 num_iters = 100
 num_warmup = 20
+
 
 def perftest(name=None):
     def decorator(func):
@@ -33,12 +34,12 @@ def perftest(name=None):
 
 
 def checkAllclose(a, b, rtol=1e-2, atol=1e-2):
-    assert torch.allclose(
-        a, b, rtol, atol), f'''torch and ck results are not close\ntorch: {a.shape}\n{a}\nck: {b.shape}\n{b}\nmax delta:{(a-b).max()}
-    delta details: {(a[(a-b)>atol]).numel()/a.numel():.1%} ({(a[(a-b)>atol]).numel()} of {a.numel()}) elements are bigger than atol: {atol}
+    isClose = torch.isclose(a, b, rtol, atol)
+    assert isClose.all(), f'''torch and ck results are not close\ntorch: {a.shape}\n{a}\nck: {b.shape}\n{b}\nmax delta:{(a-b).max()}
+    delta details: {(a[isClose]).numel()/a.numel():.1%} ({(a[isClose]).numel()} of {a.numel()}) elements {atol=} {rtol=}
           a: 
-            {a[(a-b)>atol]}
+            {a[isClose]}
           b: 
-            {b[(a-b)>atol]}
+            {b[isClose]}
       dtlta: 
-            {(a-b)[(a-b)>atol]}'''
+            {(a-b)[isClose]}'''
